# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Key Service",
    "description": "Manages API keys and BYOK (Bring Your Own Key) credentials for organizations. Handles key generation, validation, encryption, and secure storage.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "serviceKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Service-to-service key (KEY_SERVICE_API_KEY)"
      }
    },
    "schemas": {
...
      "ValidateKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "key"
        ]
      },
...
      "ListOrgKeysResponse": {
        "type": "object",
        "properties": {
          "keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgKeyItem"
            }
          }
        },
        "required": [
          "keys"
        ]
      },
      "OrgKeyItem": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateOrgKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "maskedKey": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "maskedKey",
          "message"
        ]
      },
      "CreateOrgKeyRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "minLength": 1
          },
          "apiKey": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "provider",
          "apiKey"
        ]
      },
      "DeleteOrgKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "message"
        ]
      },
...
      "DecryptPlatformKeyDirectResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "key"
        ]
      },
      "DecryptKeyResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "keySource": {
            "type": "string",
            "enum": [
              "org",
              "platform"
            ]
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "key",
          "keySource",
          "userId"
        ]
      },
      "SetKeySourceResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "keySource": {
            "type": "string",
            "enum": [
              "org",
              "platform"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "provider",
          "orgId",
          "keySource",
          "message"
        ]
      },
      "SetKeySourceRequest": {
        "type": "object",
        "properties": {
          "keySource": {
            "type": "string",
            "enum": [
              "org",
              "platform"
            ]
          }
        },
        "required": [
          "keySource"
        ]
      },
      "GetKeySourceResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string"
          },
          "orgId": {
            "type": "string"
          },
          "keySource": {
            "type": "string",
            "enum": [
              "org",
              "platform"
            ]
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "provider",
          "orgId",
          "keySource",
          "isDefault"
        ]
      },
      "ListKeySourcesResponse": {
        "type": "object",
        "properties": {
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KeySourceItem"
            }
          }
        },
        "required": [
          "sources"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/validate/keys/{provider}": {
      "get": {
        "summary": "Get decrypted org key for a provider",
        "description": "Pass the user's API key as ?key= query parameter. Requires X-Caller-Service, X-Caller-Method, and X-Caller-Path headers to identify the calling endpoint. These are used to build the provider requirements registry.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "The API key to validate (distrib.usr_*)",
              "example": "distrib.usr_abc123"
            },
            "required": true,
            "description": "The API key to validate (distrib.usr_*)",
            "name": "key",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Name of the calling service",
              "example": "apollo"
            },
            "required": true,
            "description": "Name of the calling service",
            "name": "x-caller-service",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "HTTP method of the caller's endpoint",
              "example": "POST"
            },
            "required": true,
            "description": "HTTP method of the caller's endpoint",
            "name": "x-caller-method",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Path of the caller's endpoint",
              "example": "/leads/search"
            },
            "required": true,
            "description": "Path of the caller's endpoint",
            "name": "x-caller-path",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing key parameter or required caller headers"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Key not configured"
          }
        }
      }
    },
...
    "/keys": {
      "get": {
        "summary": "List org keys for an org",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "List of org keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrgKeysResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add or update an org key",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrgKeyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Org key saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrgKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/keys/{provider}": {
      "delete": {
        "summary": "Delete an org key",
        "description": "Uses orgId from identity header (x-org-id).",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Org key deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOrgKeyResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/keys/platform/{provider}/decrypt": {
      "get": {
        "summary": "Get decrypted platform key (direct)",
        "description": "Returns the decrypted platform-level key for the given provider. No orgId/userId needed — platform keys are global. Requires X-Caller-* headers for provider requirements tracking.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Name of the calling service",
              "example": "billing"
            },
            "required": true,
            "description": "Name of the calling service",
            "name": "x-caller-service",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "HTTP method of the caller's endpoint",
              "example": "POST"
            },
            "required": true,
            "description": "HTTP method of the caller's endpoint",
            "name": "x-caller-method",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Path of the caller's endpoint",
              "example": "/billing/charge"
            },
            "required": true,
            "description": "Path of the caller's endpoint",
            "name": "x-caller-path",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted platform key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecryptPlatformKeyDirectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required caller headers"
          },
          "404": {
            "description": "Platform key not configured"
          }
        }
      }
    },
    "/keys/{provider}/decrypt": {
      "get": {
        "summary": "Get decrypted key (auto-resolves source)",
        "description": "Returns the decrypted key for a provider. Automatically resolves whether to use org or platform key based on the org's preference (default: platform). Response includes keySource indicating which was used. Requires X-Caller-* headers for provider requirements tracking. Uses orgId and userId from identity headers.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Name of the calling service",
              "example": "apollo"
            },
            "required": true,
            "description": "Name of the calling service",
            "name": "x-caller-service",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "HTTP method of the caller's endpoint",
              "example": "POST"
            },
            "required": true,
            "description": "HTTP method of the caller's endpoint",
            "name": "x-caller-method",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Path of the caller's endpoint",
              "example": "/leads/search"
            },
            "required": true,
            "description": "Path of the caller's endpoint",
            "name": "x-caller-path",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Decrypted key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecryptKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required caller headers"
          },
          "404": {
            "description": "Key not configured"
          }
        }
      }
    },
    "/keys/{provider}/source": {
      "put": {
        "summary": "Set key source preference for an org+provider",
        "description": "Set whether an org uses its own key ('org') or the platform key ('platform') for a given provider. If switching to 'org', an org key must already be stored — otherwise returns 400. Uses orgId from identity header.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetKeySourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Key source preference saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetKeySourceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or no org key stored"
          }
        }
      },
      "get": {
        "summary": "Get key source preference for an org+provider",
        "description": "Returns the current key source preference for an org+provider. If no explicit preference is set, returns 'platform' with isDefault=true. Uses orgId from identity header.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "provider",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Key source preference",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetKeySourceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/keys/sources": {
      "get": {
        "summary": "List all key source preferences for an org",
        "description": "Returns all explicit key source preferences for an org. Providers not listed default to 'platform'. Uses orgId from identity header.",
        "security": [
          {
            "serviceKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal org UUID from client-service",
              "example": "org-uuid-123"
            },
            "required": true,
            "description": "Internal org UUID from client-service",
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Internal user UUID from client-service",
              "example": "user-uuid-456"
            },
            "required": true,
            "description": "Internal user UUID from client-service",
            "name": "x-user-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Key source preferences",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListKeySourcesResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}